Basic description of text layer structure
Incomplete; there are a lot of ???s because some sleuthing remains… for now.

Having to write a whole plist library just for some formatting is probably not worth it, but when you get in the swing of it it's sort of fun to hack about and find stuff?

That said, exposing this all in an API… does not sound particularly as fun…